user local variable rather than global veriable to manage project

+++ ddd/domain/model.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Model struct {
-	ID        uint `gorm:"primarykey"`
+	ID        int64 `gorm:"primarykey"`
 	CreatedAt int64
 	UpdatedAt int64
 	DeletedAt gorm.DeletedAt `gorm:"index"`
